astrocyte-services-py/astrocyte-gateway-py/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "astrocyte-gateway-py"
version = "0.1.0"
description = "Standalone HTTP gateway for Astrocyte (FastAPI over astrocyte-py); configure via astrocyte.yaml and optional mip.yaml"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.11"
dependencies = [
    "astrocyte==0.1.0",
    "cryptography>=42",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "PyJWT>=2.8",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "httpx>=0.27",
]
pgvector = ["astrocyte-pgvector"]
otel = [
    "opentelemetry-api>=1.20",
    "opentelemetry-sdk>=1.20",
    "opentelemetry-instrumentation-fastapi>=0.46b0",
    "opentelemetry-exporter-otlp-proto-http>=1.20",
]

# `astrocyte` is not published to PyPI yet; resolve it from the sibling package when using uv.
[tool.uv.sources]
astrocyte = { path = "../../astrocyte-py", editable = true }
astrocyte-pgvector = { path = "../../adapters-storage-py/astrocyte-pgvector", editable = true }

[project.scripts]
astrocyte-gateway-py = "astrocyte_gateway.main:main"

[tool.hatch.build.targets.wheel]
packages = ["astrocyte_gateway"]

[tool.pytest.ini_options]
testpaths = ["tests"]
